requirements.txt
# Core dependencies
streamlit>=1.28.0
pillow>=9.5.0
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.28.0

# Gemini API
google-generativeai>=0.3.0

# Canvas drawing
streamlit-drawable-canvas>=0.9.0

# Image processing
opencv-python-headless>=4.8.0